+++ Familiar/Domain/Model/SpriteSheetInfo.swift
@@ -1,0 +1,9 @@
+public struct SpriteSheetInfo: Sendable {
+    public let tilesX: Int
+    public let tilesY: Int
+
+    public init(tilesX: Int, tilesY: Int) {
+        self.tilesX = tilesX
+        self.tilesY = tilesY
+    }
+}