+++ Familiar/Domain/Model/SurfaceType.swift
@@ -1,0 +1,7 @@
+public enum SurfaceType: Sendable {
+    case screenBottom
+    case screenLeft
+    case screenRight
+    case screenTop
+    case windowTop(windowID: Int)
+}