+++ Familiar/Domain/Model/WalkableSurface.swift
@@ -1,0 +1,9 @@
+public struct WalkableSurface: Sendable {
+    public let rect: Rect
+    public let type: SurfaceType
+
+    public init(rect: Rect, type: SurfaceType) {
+        self.rect = rect
+        self.type = type
+    }
+}